apps/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
requests>=2.31.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.3.0
psutil>=5.9.6
python-dotenv>=1.0.0
croniter>=1.4.1
email-validator>=2.1.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
alembic>=1.12.0
cryptography>=41.0.0
openai>=1.0.0
prometheus-client>=0.17.0 